requirements.txt
# Core Dependencies
pandas>=1.3.0
supabase>=2.0.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0

# Date/Time Handling
python-dateutil>=2.8.2

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-rest-knox==4.2.0
djangorestframework-simplejwt
Pillow

streamlit>=1.37.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
plotly>=5.22.0
textblob>=0.18.0.post0
spacy>=3.7.0
openpyxl>=3.1.2
xlrd>=2.0.1
xlsxwriter>=3.2.0
python-dotenv>=1.0.1